Address PR comments, restructure C API implementation more cleanly
