+++ cpp/tests/c_api/uniform_temporal_neighbor_sample_test.c
@@ -340,6 +340,7 @@ int generic_uniform_temporal_neighbor_sample_test(
   cugraph_sampling_set_temporal_sampling_comparison(sampling_options, temporal_sampling_comparison);
   // Temporal neighbor sampling requires disjoint sampling.
   cugraph_sampling_set_disjoint_sampling(sampling_options, TRUE);
+  cugraph_sampling_set_neighbor_selection(sampling_options, neighbor_selection);
 
   ret_code = cugraph_neighbor_sample(handle,
                                      rng_state,
@@ -352,8 +353,6 @@ int generic_uniform_temporal_neighbor_sample_test(
                                      NULL,
                                      h_fan_out_view,
                                      1,
-                                     neighbor_selection,
-                                     &temporal_sampling_comparison,
                                      sampling_options,
                                      FALSE,
                                      &result,
@@ -519,6 +518,7 @@ int test_uniform_temporal_neighbor_sample_with_labels(const cugraph_resource_han
   cugraph_sampling_set_temporal_sampling_comparison(sampling_options, MONOTONICALLY_INCREASING);
   // Temporal neighbor sampling requires disjoint sampling.
   cugraph_sampling_set_disjoint_sampling(sampling_options, TRUE);
+  cugraph_sampling_set_neighbor_selection(sampling_options, CUGRAPH_NEIGHBOR_SELECTION_RANDOM);
 
   ret_code = cugraph_homogeneous_uniform_temporal_neighbor_sample(handle,
                                                                   rng_state,
