+++ cpp/tests/sampling/temporal_neighbor_sampling_fixed_window.cpp
@@ -26,7 +26,7 @@ struct Fixed_Window_Temporal_Neighbor_Sampling_Usecase {
   cugraph::temporal_sampling_comparison_t temporal_sampling_comparison{
     cugraph::temporal_sampling_comparison_t::MONOTONICALLY_INCREASING};
   bool check_correctness{true};
-  // FIRST and LAST are expected to fast-fail until the selection primitive is implemented.
+  // LAST are expected to fast-fail until the selection primitive is implemented.
   cugraph::neighbor_selection_t neighbor_selection{cugraph::neighbor_selection_t::RANDOM};
 };
 
@@ -61,7 +61,6 @@ std::ostream& operator<<(std::ostream& os,
   os << " check_correctness=" << usecase.check_correctness << " neighbor_selection=";
   switch (usecase.neighbor_selection) {
     case cugraph::neighbor_selection_t::RANDOM: os << "RANDOM"; break;
-    case cugraph::neighbor_selection_t::FIRST: os << "FIRST"; break;
     case cugraph::neighbor_selection_t::LAST: os << "LAST"; break;
   }
   return os << "}";
@@ -289,7 +288,8 @@ class Tests_Fixed_Window_Temporal_Neighbor_Sampling
 
     if (temporal_neighbor_sampling_usecase.neighbor_selection !=
         cugraph::neighbor_selection_t::RANDOM) {
-      // FIRST and LAST are declared in the public API but are not implemented yet.  Pin the
+      sampling_flags.neighbor_selection = temporal_neighbor_sampling_usecase.neighbor_selection;
+      // LAST are declared in the public API but are not implemented yet.  Pin the
       // fast-fail so the stub cannot start silently returning edges before the selection
       // primitive replaces it.
       try {
@@ -319,10 +319,8 @@ class Tests_Fixed_Window_Temporal_Neighbor_Sampling
           raft::host_span<int32_t const>(temporal_neighbor_sampling_usecase.fanout.data(),
                                          temporal_neighbor_sampling_usecase.fanout.size()),
           std::nullopt,
-          temporal_neighbor_sampling_usecase.neighbor_selection,
-          std::make_optional(cugraph::temporal_sampling_comparison_t::FIXED_WINDOW),
           sampling_flags);
-        ADD_FAILURE() << "expected FIRST/LAST neighbor selection to fail as not implemented";
+        ADD_FAILURE() << "expected LAST neighbor selection to fail as not implemented";
       } catch (cugraph::logic_error const& e) {
         EXPECT_NE(std::string{e.what()}.find("not yet implemented"), std::string::npos)
           << "expected a not-implemented failure, got: " << e.what();
@@ -581,27 +579,16 @@ INSTANTIATE_TEST_SUITE_P(
 INSTANTIATE_TEST_SUITE_P(
   file_test_unimplemented_selection,
   Tests_Fixed_Window_Temporal_Neighbor_Sampling_File,
-  ::testing::Combine(::testing::Values(
-                       Fixed_Window_Temporal_Neighbor_Sampling_Usecase{
-                         {4, -1, 10},
-                         128,
-                         false,
-                         false,
-                         true,
-                         true,
-                         cugraph::temporal_sampling_comparison_t::FIXED_WINDOW,
-                         false,
-                         cugraph::neighbor_selection_t::FIRST},
-                       Fixed_Window_Temporal_Neighbor_Sampling_Usecase{
-                         {4, -1, 10},
-                         128,
-                         false,
-                         false,
-                         true,
-                         true,
-                         cugraph::temporal_sampling_comparison_t::FIXED_WINDOW,
-                         false,
-                         cugraph::neighbor_selection_t::LAST}),
+  ::testing::Combine(::testing::Values(Fixed_Window_Temporal_Neighbor_Sampling_Usecase{
+                       {4, -1, 10},
+                       128,
+                       false,
+                       false,
+                       true,
+                       true,
+                       cugraph::temporal_sampling_comparison_t::FIXED_WINDOW,
+                       false,
+                       cugraph::neighbor_selection_t::LAST}),
                      ::testing::Values(cugraph::test::File_Usecase("test/datasets/karate.mtx"))));
 
 CUGRAPH_TEST_PROGRAM_MAIN()
